cleanup and gamepad

+++ src/qcommon/qcommon.h
@@ -52,7 +52,7 @@ void MSG_Init( msg_t *buf, byte *data, int length );
 void MSG_InitOOB( msg_t *buf, byte *data, int length );
 void MSG_Clear( msg_t *buf );
 void *MSG_GetSpace( msg_t *buf, int length );
-void MSG_WriteData( msg_t *buf, const void *data, int length );
+void MSG_WriteData( msg_t *buf, const void *data, size_t length );
 void MSG_Bitstream( msg_t *buf );
